move cache search unit tests into a test module file
